backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud Platform
google-cloud-firestore==2.13.1
google-cloud-sql-connector==1.5.0
google-cloud-logging==3.8.0
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.12.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# ML/AI
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2

# Caching
redis==5.0.1

# Monitoring and logging
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.8
python-multipart==0.0.6